+++ common/src/main/java/net/mehvahdjukaar/moonlight/core/mixins/LoomBannerSlotMixin.java
@@ -1,0 +1,19 @@
+package net.mehvahdjukaar.moonlight.core.mixins;
+
+import net.mehvahdjukaar.moonlight.api.item.ILoomItem;
+import net.minecraft.world.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(targets = "net.minecraft.world.inventory.LoomMenu$3")
+public abstract class LoomBannerSlotMixin {
+
+    @Inject(method = "mayPlace(Lnet/minecraft/world/item/ItemStack;)Z", at = @At("HEAD"), cancellable = true)
+    private void moonlight$allowCustomBanners(ItemStack stack, CallbackInfoReturnable<Boolean> cir) {
+        if (stack.getItem() instanceof ILoomItem) {
+            cir.setReturnValue(true);
+        }
+    }
+}